mentorship/pyproject.toml
[tool.poetry]
name = "mentorship"
version = "0.1.0"
description = ""
authors = ["s.smirnov <s.smirnov@tm-industries.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.2"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
celery = "^5.4.0"
redis = "^5.2.0"
gunicorn = "^23.0.0"
pytest = "^8.3.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
